MBS-9422 partially deal with duplication in tests
