+++ subprojects/gradle/cd/src/test/kotlin/com/avito/ci/steps/FastCheckChangesInAndroidTestInAppA.kt
@@ -1,6 +1,7 @@
 package com.avito.ci.steps
 
 import com.avito.ci.assertAffectedModules
+import com.avito.ci.generateProjectWithImpactAnalysis
 import com.avito.test.gradle.TestProjectGenerator
 import com.avito.test.gradle.TestResult
 import com.avito.test.gradle.ciRun
@@ -14,7 +15,7 @@ import org.junit.jupiter.api.io.TempDir
 import java.io.File
 import java.nio.file.Path
 
-class FastCheck_ChangesInAndroidTestInAppA {
+class FastCheckChangesInAndroidTestInAppA {
 
     private lateinit var projectDir: File
