+++ subprojects/gradle/cd/src/test/kotlin/com/avito/ci/steps/FastCheckNoChanges.kt
@@ -1,25 +1,22 @@
 package com.avito.ci.steps
 
 import com.avito.ci.assertAffectedModules
-import com.avito.test.gradle.TestProjectGenerator
+import com.avito.ci.generateProjectWithImpactAnalysis
 import com.avito.test.gradle.TestResult
 import com.avito.test.gradle.ciRun
-import com.avito.test.gradle.commit
-import com.avito.test.gradle.file
 import com.avito.test.gradle.git
-import com.avito.test.gradle.mutate
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.io.TempDir
 import java.io.File
 import java.nio.file.Path
 
-class FastCheck_OneAppBChanges {
+class FastCheckNoChanges {
 
     private lateinit var projectDir: File
 
     private val targetBranch = "develop"
-    private val sourceBranch = "changes-in-the-${TestProjectGenerator.appB}"
+    private val sourceBranch = "no-changes"
 
     @BeforeEach
     fun setup(@TempDir tempDir: Path) {
@@ -31,16 +28,14 @@ class FastCheck_OneAppBChanges {
             git("checkout -b $targetBranch")
 
             git("checkout -b $sourceBranch $targetBranch")
-            file("${TestProjectGenerator.appB}/src/main/kotlin/SomeClass.kt").mutate()
-            commit()
         }
     }
 
     @Test
-    fun `fastCheck triggers build only in appA`() {
+    fun `fastCheck - does not trigger any of assembleDebug tasks - thre is no changes`() {
         val result = runTask("fastCheck")
 
-        result.assertAffectedModules("packageDebug", setOf(":${TestProjectGenerator.appB}"))
+        result.assertAffectedModules("assemble", emptySet())
     }
 
     private fun runTask(taskName: String): TestResult =
